+++ src/app/dishes/create/dish-ingredient.tsx
@@ -77,7 +77,24 @@ export function DishIngredient({ ingredient }: DishIngredientProps) {
         <li>
             <InputNumber label={'Ilość'} value={amount} setValue={setAmountValue} width={'150px'} shouldHaveMargin={true} customError={customAmountError} />
             <InputSelect id={`${ingredient.id}-unit`} label={'Jednostka'} options={unitOptions} selectedValue={unit} setSelectedValue={setUnitValue} width={'150px'} shouldHaveMargin={true} customError={customUnitError} />
-            <div className={styles['ingredient-name']} style={{ marginTop: '26px', alignItems: 'flex-start' }}>{ingredient.data.pl}</div>
+            <div className={styles['ingredient-name']} style={{ marginTop: '26px', alignItems: 'flex-start' }}>
+                {ingredient.data.imageUrl
+                    ? <img src={`https://img.spoonacular.com/ingredients_250x250/${ingredient.data.imageUrl}`}
+                        alt="Ingredient image icon"
+                        width={32}
+                        height={32}
+                        style={{ borderRadius: '50%', marginRight: '0.3rem' }}
+                    />
+                    : <img src="/ingredient.png"
+                        alt="Generic ingredient image icon"
+                        width={32}
+                        data-a-href="https://www.flaticon.com/free-icons/ingredients"
+                        data-a-title="ingredients icons"
+                        data-a-text="Ingredients icons created by Flat Icons - Flaticon"
+                    />
+                }
+                {ingredient.data.pl}
+            </div>
             <RemoveButton label={'Usuń'} onClick={onRemoveIngredient} customStyle={{ marginTop: '26px', alignItems: 'flex-start' }} />
         </li>
     );