+++ src/app/dishes/create/page.tsx
@@ -8,7 +8,7 @@ import path from 'path';
 import fs from 'fs';
 
 export default function CreateDishPage() {
-    const ingredients = fetchIngredients();
+    const ingredients = getMergedIngredients();
 
     return (
         <>
@@ -29,4 +29,18 @@ export function fetchIngredients(): IngredientDataValue[] {
     const ingredients: IngredientDataValue[] = Object.entries(json).map(el => el[1]);
 
     return ingredients;
+}
+
+export function getMergedIngredients(): IngredientDataValue[] {
+    const categories = ['breads', 'cereal-products', 'dairy-and-eggs', 'fish-and-seafood', 'fruits', 'meats', 'mushrooms', 'oils-and-fats', 'pasta', 'seeds-and-nuts', 'spices', 'vegetables'];
+    const ingredientsArray = categories.map(category => {
+        const filePath = path.join(process.cwd(), `public/data/ingredients/${category}.json`);
+        const ingredientsData = fs.readFileSync(filePath, 'utf-8');
+        const json: IngredientData = JSON.parse(ingredientsData);
+        const ingredients: IngredientDataValue[] = Object.entries(json).map(el => el[1]);
+
+        return ingredients;
+    });
+
+    return ingredientsArray.flat();
 }